moved Layout in container folder and cleaned up extra lines
