+++ portfolio/components/Accomplishments/Accomplishments.js
@@ -1,8 +1,6 @@
 import React from "react";
-
 import { Section } from "../../styles/GlobalComponents";
 import { Box, Boxes, BoxNum, BoxText } from "./AccomplishmentsStyles";
-
 import { AccomplishmentsData } from "../../constants/constants";
 
 const Accomplishments = () => (
