+++ portfolio/components/Hero/Hero.js
@@ -1,5 +1,4 @@
 import React from "react";
-
 import {
   Section,
   SectionText,
