+++ portfolio/components/Layout/Layout.js
@@ -1,10 +1,9 @@
 import React from "react";
-
-import Footer from "../components/Footer/Footer";
-import Header from "../components/Header/Header";
+import Footer from "../Footer/Footer";
+import Header from "../Header/Header";
 import { Container } from "./LayoutStyles";
 
-export const Layout = ({ children }) => {
+const Layout = ({ children }) => {
   return (
     <Container>
       <Header />
@@ -13,3 +12,5 @@ export const Layout = ({ children }) => {
     </Container>
   );
 };
+
+export default Layout;
