+++ portfolio/components/Projects/Projects.js
@@ -1,5 +1,4 @@
 import React from "react";
-
 import {
   BlogCard,
   CardInfo,
