+++ portfolio/pages/index.js
@@ -3,7 +3,7 @@ import Hero from "../components/Hero/Hero";
 import Projects from "../components/Projects/Projects";
 import Skills from "../components/Skills/Skills";
 import Timeline from "../components/TimeLine/TimeLine";
-import { Layout } from "../layout/Layout";
+import Layout from "../components/Layout/Layout";
 import React from "react";
 
 const Home = () => {
